+++ app/jobs/agent_check_job.rb
@@ -1,3 +1,4 @@
+# TODO: move into runner?
 class AgentCheckJob < ActiveJob::Base
   # Given an Agent id, load the Agent, call #check on it, and then save it with an updated `last_check_at` timestamp.
   # rubocop:disable Style/RescueStandardError