+++ app/jobs/agent_receive_job.rb
@@ -1,3 +1,4 @@
+# Move into runner?
 class AgentReceiveJob < ActiveJob::Base
   # Given an Agent id and an array of Message ids, load the Agent, call #receive on it with the Message objects, and then
   # save it with an updated `last_receive_at` timestamp.