+++ app/jobs/agent_run_schedule_job.rb
@@ -1,7 +1,0 @@
-class AgentRunScheduleJob < ActiveJob::Base
-  queue_as :default
-
-  def perform(time)
-    Agent.run_schedule(time)
-  end
-end