+++ db/migrate/20200305052635_create_agent_locks.rb
@@ -1,0 +1,9 @@
+class CreateAgentLocks < ActiveRecord::Migration[6.0]
+  def change
+    create_table :agent_locks do |t|
+      t.integer 'agent_id'
+      t.datetime 'locked_at'
+      t.index ['agent_id'], name: 'index_agent_locks_on_agent_id', unique: true
+    end
+  end
+end