+++ db/migrate/20200305160036_create_agent_scheduled_table.rb
@@ -1,0 +1,8 @@
+class CreateAgentScheduledTable < ActiveRecord::Migration[6.0]
+  def change
+    create_table :agent_scheduled do |t|
+      t.integer "agent_id"
+      t.index ["agent_id"], name: "index_agent_scheduled_on_agent_id", unique: true
+    end
+  end
+end